test: add and use evalProposal
