+++ packages/boot/test/bootstrapTests/test-zcf-upgrade.ts
@@ -88,30 +88,9 @@ test.before(async t => {
 test.after.always(t => t.context.shutdown?.());
 
 test('run restart-vats proposal', async t => {
-  const { controller, buildProposal, zoeDriver } = t.context;
+  const { controller, buildProposal, evalProposal, zoeDriver } = t.context;
   const { EV } = t.context.runUtils;
 
-  const buildAndExecuteProposal = async packageSpec => {
-    const proposal = await buildProposal(packageSpec);
-
-    for await (const bundle of proposal.bundles) {
-      await controller.validateAndInstallBundle(bundle);
-    }
-
-    t.log('installed', proposal.bundles.length, 'bundles');
-
-    t.log('launching proposal');
-    const bridgeMessage = {
-      type: 'CORE_EVAL',
-      evals: proposal.evals,
-    };
-
-    t.log({ bridgeMessage });
-    const coreEvalBridgeHandler: ERef<BridgeHandler> = await EV.vat(
-      'bootstrap',
-    ).consumeItem('coreEvalBridgeHandler');
-    await EV(coreEvalBridgeHandler).fromBridge(bridgeMessage);
-  };
   const source = `${dirname}/${ZCF_PROBE_SRC}`;
   const zcfProbeBundle = await bundleSource(source);
   await controller.validateAndInstallBundle(zcfProbeBundle);
@@ -139,7 +118,9 @@ test('run restart-vats proposal', async t => {
 
   t.log('building proposal');
   // /////// Upgrading ////////////////////////////////
-  await buildAndExecuteProposal('@agoric/builders/scripts/vats/replace-zoe.js');
+  await evalProposal(
+    buildProposal('@agoric/builders/scripts/vats/replace-zoe.js'),
+  );
 
   t.log('upgrade zoe&zcf proposal executed');
   await zoeDriver.upgradeProbe(zcfProbeBundle);
